# Social Media Platform/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Social Media Platform API
  version: 1.0.0
  description: REST API documentation for the Selamat social media platform.
servers:
  - url: http://localhost:8080/api
    description: Local development server
  - url: https://api.selamat.example.com/api
    description: Production server placeholder
...
paths:
  /auth/signup:
    post:
      tags: [Authentication]
      summary: Register a new user account
      operationId: signUp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpRequest'
      responses:
        '201':
          description: User account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      tags: [Authentication]
      summary: Login with username or email and password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh an expired access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [Authentication]
      summary: Logout the current user
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Logout successful
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/forgot-password:
    post:
      tags: [Authentication]
      summary: Request a password reset message
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Password reset message returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordResetMessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/reset-password:
    post:
      tags: [Authentication]
      summary: Reset a password using a reset token
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset message returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordResetMessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /posts/{postId}/comments:
    get:
      tags: [Comments]
      summary: List comments for a post
      operationId: listPostComments
      parameters:
        - $ref: '#/components/parameters/PostId'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Comments returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentPage'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Comments]
      summary: Create a comment on a post
      operationId: createComment
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PostId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
      responses:
        '201':
          description: Comment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /admin/reports:
    get:
      tags: [Admin]
      summary: List reports for moderation review
      operationId: adminListReports
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ReportStatus'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Reports returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/reports/summary:
    get:
      tags: [Admin]
      summary: Get moderation report dashboard summary
      operationId: adminGetReportSummary
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Report summary returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminReportSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/reports/{reportId}:
    get:
      tags: [Admin]
      summary: Get one moderation report
      operationId: adminGetReportById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ReportId'
      responses:
        '200':
          description: Report returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/reports/{reportId}/target:
    get:
      tags: [Admin]
      summary: Get the target of a moderation report
      operationId: adminGetReportTarget
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ReportId'
      responses:
        '200':
          description: Report target returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportTarget'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/reports/{reportId}/resolve:
    patch:
      tags: [Admin]
      summary: Resolve a moderation report
      operationId: adminResolveReport
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ReportId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveReportRequest'
      responses:
        '200':
          description: Report resolved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/moderation/posts/{postId}/hide:
    patch:
      tags: [Admin]
      summary: Hide a post
      operationId: adminHidePost
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PostId'
      responses:
        '204':
          description: Post hidden
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/moderation/comments/{commentId}/hide:
    patch:
      tags: [Admin]
      summary: Hide a comment
      operationId: adminHideComment
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CommentId'
      responses:
        '204':
          description: Comment hidden
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/moderation/messages/{messageId}/delete:
    patch:
      tags: [Admin]
      summary: Delete a message
      operationId: adminDeleteMessage
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MessageId'
      responses:
        '204':
          description: Message deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/moderation/media/{mediaId}/hide:
    patch:
      tags: [Admin]
      summary: Hide media
      operationId: adminHideMedia
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MediaId'
      responses:
        '204':
          description: Media hidden
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/moderation/users/{userId}/suspend:
    patch:
      tags: [Admin]
      summary: Suspend a user
      operationId: adminSuspendUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: User suspended
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PostId:
      name: postId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    CommentId:
      name: commentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    MessageId:
      name: messageId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    MediaId:
      name: mediaId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    ReportId:
      name: reportId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
        format: date-time
      description: ISO-8601 cursor returned from a previous paginated response.
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication is required or the token is invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: The authenticated user does not have permission
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Resource conflict, such as duplicate username or email
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnprocessableEntity:
      description: Request was valid JSON but failed business validation or moderation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    SignUpRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: tinsae
        email:
          type: string
          format: email
          example: tinsae@example.com
        password:
          type: string
          format: password
          minLength: 8
          example: StrongPassword123
    LoginRequest:
      type: object
      required: [identifier, password]
      properties:
        identifier:
          type: string
          description: Username or email address.
          example: tinsae
        password:
          type: string
          format: password
          example: StrongPassword123
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          maxLength: 255
    ResetPasswordRequest:
      type: object
      required: [token, newPassword]
      properties:
        token:
          type: string
        newPassword:
          type: string
          format: password
          minLength: 8
          maxLength: 100
    PasswordResetMessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    AuthResponse:
      type: object
      required: [accessToken, refreshToken, tokenType, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      required: [id, username, email, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        status:
          $ref: '#/components/schemas/UserStatus'
        roles:
          type: array
          items:
            type: string
          example: [USER]
        createdAt:
          type: string
          format: date-time
...
    CreateCommentRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          maxLength: 2000
        parentCommentId:
          type: string
          format: uuid
          nullable: true
    Comment:
      type: object
      required: [id, postId, author, content, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        postId:
          type: string
          format: uuid
        parentCommentId:
          type: string
          format: uuid
          nullable: true
        author:
          $ref: '#/components/schemas/UserSummary'
        content:
          type: string
        status:
          $ref: '#/components/schemas/CommentStatus'
        reactionCount:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
...
    ResolveReportRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/ReportStatus'
        resolutionNote:
          type: string
          maxLength: 1000
    Report:
      type: object
      required: [id, reporter, targetType, targetId, reason, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        reporter:
          $ref: '#/components/schemas/UserSummary'
        resolver:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/UserSummary'
        targetType:
          $ref: '#/components/schemas/ReportTargetType'
        targetId:
          type: string
          format: uuid
        reason:
          type: string
        details:
          type: string
        resolutionNote:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/ReportStatus'
        createdAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
          nullable: true
    AdminReportSummary:
      type: object
      required:
        - openReports
        - underReviewReports
        - resolvedReports
        - rejectedReports
        - userReports
        - contentReports
        - mediaReports
      properties:
        openReports:
          type: integer
          format: int64
        underReviewReports:
          type: integer
          format: int64
        resolvedReports:
          type: integer
          format: int64
        rejectedReports:
          type: integer
          format: int64
        userReports:
          type: integer
          format: int64
        contentReports:
          type: integer
          format: int64
        mediaReports:
          type: integer
          format: int64
    ReportTarget:
      type: object
      required: [reportId, targetType, targetId]
      properties:
        reportId:
          type: string
          format: uuid
        targetType:
          $ref: '#/components/schemas/ReportTargetType'
        targetId:
          type: string
          format: uuid
        user:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/UserSummary'
        email:
          type: string
          format: email
          nullable: true
        userStatus:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/UserStatus'
        bio:
          type: string
          nullable: true
        author:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/UserSummary'
        sender:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/UserSummary'
        recipient:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/UserSummary'
        uploader:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/UserSummary'
        postId:
          type: string
          format: uuid
          nullable: true
        content:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
        visibility:
          type: string
          nullable: true
        commentCount:
          type: integer
          format: int64
          nullable: true
        reactionCount:
          type: integer
          format: int64
          nullable: true
        url:
          type: string
          format: uri
          nullable: true
        mediaType:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/MediaType'
        mimeType:
          type: string
          nullable: true
        sizeBytes:
          type: integer
          format: int64
          nullable: true
        altText:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
          nullable: true
        media:
          type: array
          items:
            $ref: '#/components/schemas/MediaPreview'
    MediaPreview:
      type: object
      required: [id, url, mediaType, mimeType, sizeBytes]
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        mediaType:
          $ref: '#/components/schemas/MediaType'
        mimeType:
          type: string
        sizeBytes:
          type: integer
          format: int64
        altText:
          type: string
          nullable: true
    UserSummary:
      type: object
      required: [id, username]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
          format: uri
          nullable: true
...
    CommentPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
        nextCursor:
          type: string
          nullable: true
...
    ReportPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Report'
        nextCursor:
          type: string
          nullable: true
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Request validation failed.
        details:
          type: array
          items:
            type: string
    UserStatus:
      type: string
      enum: [ACTIVE, SUSPENDED, DELETED]
...
    MediaType:
      type: string
      enum: [IMAGE, VIDEO, AUDIO, DOCUMENT]
    CommentStatus:
      type: string
      enum: [VISIBLE, HIDDEN, DELETED]
...
    ReportTargetType:
      type: string
      enum: [USER, POST, COMMENT, MESSAGE, MEDIA]
    ReportStatus:
      type: string
      enum: [OPEN, UNDER_REVIEW, RESOLVED, REJECTED]