# Social Media Platform/src/main/java/com/tinsae/socialmediaplatform/auth/dto/RefreshTokenRequest.java
package com.tinsae.socialmediaplatform.auth.dto;

import jakarta.validation.constraints.NotBlank;

public record RefreshTokenRequest(
        @NotBlank(message = "Refresh token is required.")
        String refreshToken
) {
}